requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blessed==1.20.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
dotenv==0.9.9
google-auth==2.38.0
google-genai==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
readchar==4.2.1
requests==2.32.3
revel==0.9.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
u==1.0
urllib3==2.3.0
wcwidth==0.2.13
websockets==15.0.1
